test: get delegation
